+++ src/app/(withLayout)/category/[id]/[bookDetails]/page.jsx
@@ -1,6 +1,7 @@
 import BookDetails from "@/components/Home/healthAndFitness/bookDetails/BookDetails";
 
 const page = async ({ params }) => {
+    console.log(params.id)
     return (
         <BookDetails params={params}></BookDetails>
     );
