--- src/app/(withLayout)/category/article/page.jsx
@@ -1,12 +1,0 @@
-
-
-const page = ({params}) => {
-    console.log(params)
-    return (
-        <div>
-            jjjj
-        </div>
-    );
-};
-
-export default page;
